refactor: split server into modules
