+++ server/userStore.ts
@@ -1,0 +1,9 @@
+export interface UserRecord {
+  id: string;
+  password: string;
+}
+
+export const users: Record<string, UserRecord> = {
+  user: { id: 'user', password: 'pass' },
+  test: { id: 'test', password: 'testpass' },
+};
